Create update_recurring_invoice_selling_price
This function should be triggered when an item has its selling price changed. 
It then runs through all recurring invoices and updates the selling price on those for the item. If you have more than 2,000 recurring invoices, you may have to increase the pages dimension to cater for this
When testing, the workflow logs will show you it is scheduled, then processed (with a success update).